backend/requirements.txt
Django>=4.0
djangorestframework
mysqlclient
pandas
openpyxl
django-cors-headers

